requirements.txt
streamlit
fastai
torch
pillow